Generate HelpMojo for turmeric-maven-plugin
This update to the pom makes use of the maven-plugin-plugin.  It executes
the helpmojo to generate a HelpMojo class for use with the
turmeric-maven-plugin.   Just run turmeric:help to get the necessary
information. It will print out help for each of the goals or you can
specify a specific goal to get help on that particular goal.